requirements.txt
aiohttp==3.8.5
aioselenium==0.0.1
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
beautifulsoup4==4.12.2
boto3==1.28.22
botocore==1.31.22
certifi==2023.7.22
charset-normalizer==3.2.0
exceptiongroup==1.1.2
frozenlist==1.4.0
h11==0.14.0
idna==3.4
jmespath==1.0.1
multidict==6.0.4
numpy==1.25.2
outcome==1.2.0
packaging==23.1
pandas==2.0.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
s3transfer==0.6.1
selenium==4.11.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
trio==0.22.2
trio-websocket==0.10.3
tzdata==2023.3
urllib3==1.26.16
wsproto==1.2.0
yarl==1.9.2
